Refactor workflow power unit and rename related classes
Changes were made to the Workflow and Serial Services to accommodate the power unit conversion from Watt-hours to NanoJoules. The unit conversion is now done directly in the method RunWorkflowProcedure for more accurate energy calculation. The WorkflowController was deleted; StreamController was renamed to DetectionController to better represent its functionality. The StreamInstructController and SerialServiceTests classes were updated accordingly.

+++ src/Sprinti/Serial/SerialService.cs
@@ -9,7 +9,7 @@ public interface ISerialService
     Task<FinishedResponse> SendCommand(FinishCommand command, CancellationToken cancellationToken);
     Task<string> SendRawCommand(string command, CancellationToken stoppingToken);
 
-    Task<int> RunWorkflowProcedure(IEnumerable<ISerialCommand> instructions, CancellationToken cancellationToken);
+    Task<double> RunWorkflowProcedure(IEnumerable<ISerialCommand> instructions, CancellationToken cancellationToken);
     Task RunStartProcedure(CancellationToken cancellationToken);
 }
 
@@ -47,7 +47,7 @@ public class SerialService(
         return await readTask.WaitAsync(Timeout, stoppingToken);
     }
 
-    public async Task<int> RunWorkflowProcedure(IEnumerable<ISerialCommand> instructions,
+    public async Task<double> RunWorkflowProcedure(IEnumerable<ISerialCommand> instructions,
         CancellationToken cancellationToken)
     {
         logger.LogInformation("Instructions: {instructions}", instructions);
